+++ addons/mail/__manifest__.py
@@ -10,12 +10,12 @@
     'website': 'https://www.odoo.com/app/discuss',
     'depends': ['base', 'base_setup', 'bus', 'web_tour'],
     'data': [
-        'wizard/invite_view.xml',
-        'wizard/mail_blacklist_remove_view.xml',
-        'wizard/mail_compose_message_view.xml',
+        'wizard/mail_blacklist_remove_views.xml',
+        'wizard/mail_compose_message_views.xml',
         'wizard/mail_resend_cancel_views.xml',
         'wizard/mail_resend_message_views.xml',
         'wizard/mail_template_preview_views.xml',
+        'wizard/mail_wizard_invite_views.xml',
         'views/mail_message_subtype_views.xml',
         'views/mail_tracking_views.xml',
         'views/mail_notification_views.xml',
@@ -38,7 +38,7 @@
         'views/mail_alias_views.xml',
         'views/res_users_views.xml',
         'views/mail_template_views.xml',
-        'views/ir_actions_views.xml',
+        'views/ir_actions_server_views.xml',
         'views/ir_model_views.xml',
         'views/res_partner_views.xml',
         'views/mail_blacklist_views.xml',
