+++ addons/mail/controllers/__init__.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*
-from . import main
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
 from . import bus
+from . import discuss
 from . import home
+from . import mail
