+++ addons/mail/controllers/mail.py
@@ -1,10 +1,7 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-import base64
 import logging
-import werkzeug.utils
-import werkzeug.wrappers
 
 from werkzeug.urls import url_encode
 
@@ -125,63 +122,6 @@ class MailController(http.Controller):
         url = '/web?#%s' % url_encode(url_params)
         return request.redirect(url)
 
-    @http.route('/mail/read_followers', type='json', auth='user')
-    def read_followers(self, res_model, res_id):
-        request.env['mail.followers'].check_access_rights("read")
-        request.env[res_model].check_access_rights("read")
-        request.env[res_model].browse(res_id).check_access_rule("read")
-        follower_recs = request.env['mail.followers'].search([('res_model', '=', res_model), ('res_id', '=', res_id)])
-
-        followers = []
-        follower_id = None
-        for follower in follower_recs:
-            if follower.partner_id == request.env.user.partner_id:
-                follower_id = follower.id
-            followers.append({
-                'id': follower.id,
-                'partner_id': follower.partner_id.id,
-                'name': follower.name,
-                'display_name': follower.display_name,
-                'email': follower.email,
-                'is_active': follower.is_active,
-                # When editing the followers, the "pencil" icon that leads to the edition of subtypes
-                # should be always be displayed and not only when "debug" mode is activated.
-                'is_editable': True
-            })
-        return {
-            'followers': followers,
-            'subtypes': self.read_subscription_data(follower_id) if follower_id else None
-        }
-
-    @http.route('/mail/read_subscription_data', type='json', auth='user')
-    def read_subscription_data(self, follower_id):
-        """ Computes:
-            - message_subtype_data: data about document subtypes: which are
-                available, which are followed if any """
-        request.env['mail.followers'].check_access_rights("read")
-        follower = request.env['mail.followers'].sudo().browse(follower_id)
-        follower.ensure_one()
-        request.env[follower.res_model].check_access_rights("read")
-        request.env[follower.res_model].browse(follower.res_id).check_access_rule("read")
-
-        # find current model subtypes, add them to a dictionary
-        subtypes = request.env['mail.message.subtype'].search([
-            '&', ('hidden', '=', False),
-            '|', ('res_model', '=', follower.res_model), ('res_model', '=', False)])
-        followed_subtypes_ids = set(follower.subtype_ids.ids)
-        subtypes_list = [{
-            'name': subtype.name,
-            'res_model': subtype.res_model,
-            'sequence': subtype.sequence,
-            'default': subtype.default,
-            'internal': subtype.internal,
-            'followed': subtype.id in followed_subtypes_ids,
-            'parent_model': subtype.parent_id.res_model,
-            'id': subtype.id
-        } for subtype in subtypes]
-        return sorted(subtypes_list,
-                      key=lambda it: (it['parent_model'] or '', it['res_model'] or '', it['internal'], it['sequence']))
-
     @http.route('/mail/view', type='http', auth='public')
     def mail_action_view(self, model=None, res_id=None, access_token=None, **kwargs):
         """ Generic access point from notification emails. The heuristic to
@@ -222,56 +162,3 @@ class MailController(http.Controller):
             except Exception:
                 return self._redirect_to_messaging()
         return redirect
-
-    @http.route('/mail/<string:res_model>/<int:res_id>/avatar/<int:partner_id>', type='http', auth='public')
-    def avatar(self, res_model, res_id, partner_id):
-        headers = [('Content-Type', 'image/png')]
-        status = 200
-        content = 'R0lGODlhAQABAIABAP///wAAACH5BAEKAAEALAAAAAABAAEAAAICTAEAOw=='  # default image is one white pixel
-        if res_model in request.env:
-            try:
-                # if the current user has access to the document, get the partner avatar as sudo()
-                request.env[res_model].browse(res_id).check_access_rule('read')
-                if partner_id in request.env[res_model].browse(res_id).sudo().exists().message_ids.mapped('author_id').ids:
-                    status, headers, _content = request.env['ir.http'].sudo().binary_content(
-                        model='res.partner', id=partner_id, field='avatar_128', default_mimetype='image/png')
-                    # binary content return an empty string and not a placeholder if obj[field] is False
-                    if _content != '':
-                        content = _content
-                    if status == 304:
-                        return werkzeug.wrappers.Response(status=304)
-            except AccessError:
-                pass
-        image_base64 = base64.b64decode(content)
-        headers.append(('Content-Length', len(image_base64)))
-        response = request.make_response(image_base64, headers)
-        response.status = str(status)
-        return response
-
-    @http.route('/mail/needaction', type='json', auth='user')
-    def needaction(self):
-        return request.env.user.partner_id._get_needaction_count()
-
-    @http.route('/mail/init_messaging', type='json', auth='user')
-    def mail_init_messaging(self):
-        return request.env.user._init_messaging()
-
-    @http.route('/mail/get_partner_info', type='json', auth='user')
-    def message_partner_info_from_emails(self, model, res_ids, emails, link_mail=False):
-        records = request.env[model].browse(res_ids)
-        try:
-            records.check_access_rule('read')
-            records.check_access_rights('read')
-        except:
-            return []
-        return records._message_partner_info_from_emails(emails, link_mail=link_mail)
-
-    @http.route('/mail/get_suggested_recipients', type='json', auth='user')
-    def message_get_suggested_recipients(self, model, res_ids):
-        records = request.env[model].browse(res_ids)
-        try:
-            records.check_access_rule('read')
-            records.check_access_rights('read')
-        except:
-            return {}
-        return records._message_get_suggested_recipients()
