+++ addons/mail/views/ir_actions_server_views.xml
