+++ addons/mail/wizard/__init__.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import invite
+from . import base_module_uninstall
+from . import base_partner_merge_automatic_wizard
 from . import mail_blacklist_remove
 from . import mail_compose_message
 from . import mail_resend_cancel
 from . import mail_resend_message
 from . import mail_template_preview
-from . import base_module_uninstall
-from . import base_partner_merge
+from . import mail_wizard_invite
