+++ addons/mail/wizard/mail_compose_message_views.xml
