+++ addons/mail/wizard/mail_wizard_invite.py
