+++ addons/mail/wizard/mail_wizard_invite_views.xml
