+++ addons/portal/controllers/mail.py
@@ -8,7 +8,7 @@ from odoo import http, _
 from odoo.http import request
 from odoo.osv import expression
 from odoo.tools import consteq, plaintext2html
-from odoo.addons.mail.controllers import main
+from odoo.addons.mail.controllers import mail
 from odoo.addons.portal.controllers.portal import CustomerPortal
 from odoo.exceptions import AccessError, MissingError, UserError
 
@@ -205,7 +205,7 @@ class PortalChatter(http.Controller):
         return message.is_internal
 
 
-class MailController(main.MailController):
+class MailController(mail.MailController):
 
     @classmethod
     def _redirect_to_record(cls, model, res_id, access_token=None, **kwargs):
